+++ spec/features/elements_feature_spec.rb
@@ -1,4 +1,3 @@
-# coding: utf-8
 require 'rails_helper'
 require_relative 'concerns/index_feature_helper'